[Refactor] protocol 확장 메서드 정의
