+++ MyMemory/MyMemory/View/Authentication/ProfileMemoListCell.swift
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct ProfileMemoListCell<ViewModel: ProfileViewModelProtocal>: View {
+struct ProfileMemoListCell<ViewModel: ProfileViewModelProtocol>: View {
     
     @Binding var memo: Memo
     @ObservedObject var viewModel: ViewModel
